Move model schema to model directory
