+++ src/events/attendance/voiceStateUpdate.ts
@@ -1,6 +1,6 @@
 import { Events } from 'discord.js';
 import { Event } from '../../Classes/index.js';
-import { VoiceSession } from '../../features/attendance/index.js';
+import { VoiceSession } from '../../models/attendance/index.js';
 import dbConnect from "../../util/libmongo.js";
 
 /** Records when people enter and leave channels */
