+++ src/models/attendance/index.ts
@@ -1,0 +1,4 @@
+export { ScheduledEvent } from './ScheduledEvent.js';
+export { ScheduledEventInterest } from './ScheduledEventInterest.js';
+export { VoiceSession } from './VoiceSession.js';
+
